1.4.0
Switch build system to Hatchling and reformatted project configurations.

Introduce two Django CBV mixins to speed up development for projects using this library.

```AddWebSocketRouteToContextMixin``` this mixin can vbe added to CBV views that need to connect to and send message to the server via Websockets. It allows you to specific the ```route_key``` to use to idnetify the WebSocket route you created (defaulting to the "default" route), as well as adding a class property and context key of "channel_name" that can be used within the HTML templates to connect to the channel intended to be used by the view.

```AppChannelWebSocketMixin``` this mixin extends the first with the intention of being used by CBVs where the both the ```route_key``` and ```channel_name``` are the same and *BOTH* are the same names as the app. This is useful in the design paradigm where you are going to have one WebSocket endpoint per app handling the server-side requests. This way you can create the route key to match the Django app name and then use this mixin to extend the CBVs that render the client facing HTML.